New Overview Page stage 3
Split into separate component for each overview box
