+++ frontend/src/app/components/overview/buckets-overview/buckets-overview.less
@@ -1,0 +1,14 @@
+buckets-overview {
+    display: block;
+
+    /*
+    https://unsplash.com/license
+    https://unsplash.com/collections/136301/space?photo=NORa8-4ohA0
+    */
+    background: url('@{assets}/stars.jpg');
+    background-size: cover;
+
+    .duration-chooser .toggle-btn {
+        min-width: 90px;
+    }
+}
