+++ frontend/src/app/components/overview/cloud-overview/cloud-overview.less
@@ -1,0 +1,2 @@
+cloud-overview {
+}
