--- frontend/src/app/components/overview/cloud-summary/cloud-summary.html
@@ -1,28 +1,0 @@
-<div class="row push-prev">
-    <div class="column content-middle greedy">
-        <svg-icon class="icon-medium push-next"
-            params="name: awsResourceIcon">
-        </svg-icon>
-        <span class="remark highlight" data-bind="
-            text: awsResourceCount
-        "></span>
-    </div>
-
-    <div class="column content-middle greedy">
-        <svg-icon class="icon-medium push-next"
-            params="name: azureResourceIcon">
-        </svg-icon>
-        <span class="remark highlight" data-bind="
-            text: azureResourceCount
-        "></span>
-    </div>
-
-    <div class="column content-middle greedy">
-        <svg-icon  class="icon-medium push-next"
-            params="name: genericResourceIcon">
-        </svg-icon>
-        <span class="remark highlight" data-bind="
-            text: genericResourceCount
-        "></span>
-    </div>
-</div>
