--- frontend/src/app/components/overview/cloud-summary/cloud-summary.less
@@ -1,3 +1,0 @@
-cloud-summary {
-    .column();
-}
