+++ frontend/src/app/components/overview/cluster-overview/cluster-overview.html
@@ -1,0 +1,10 @@
+<div class="column pad borders greedy">
+    <h2 class="heading3 row content-middle">
+        <span class="greedy">Cluster</span>
+        <a class="link" data-bind="
+            text: serverCount,
+            href: { route: 'cluster' }
+        "></a>
+    </h2>
+    <hr>
+</div>
