+++ frontend/src/app/components/overview/cluster-overview/cluster-overview.less
@@ -1,0 +1,2 @@
+cluster-overview {
+}
