--- frontend/src/app/components/overview/nodes-summary/nodes-summary.less
@@ -1,7 +1,0 @@
-nodes-summary {
-    .column();
-
-    .nodes-chart {
-        margin: auto;
-    }
-}
