+++ frontend/src/app/components/overview/overview-panel/overview-panel.less
@@ -7,15 +7,6 @@
         margin-top: 50px;
     }
 
-    .stars-bg {
-        /*
-        https://unsplash.com/license
-        https://unsplash.com/collections/136301/space?photo=NORa8-4ohA0
-        */
-        background: url('@{assets}/stars.jpg');
-        background-size: cover;
-    }
-
     .overview-btn {
         font-size: @font-size4;
         width :100%;
