+++ frontend/src/app/components/overview/pools-overview/pools-overview.less
@@ -1,0 +1,15 @@
+pools-overview {
+    display: block;
+
+    /*
+    https://unsplash.com/license
+    https://unsplash.com/collections/136301/space?photo=NORa8-4ohA0
+    */
+    background: url('@{assets}/stars.jpg');
+    background-size: cover;
+
+
+    .nodes-chart {
+        margin: auto;
+    }
+}
