--- frontend/src/app/components/overview/storage-usage-summary/storage-usage-summary.html
@@ -1,15 +1,0 @@
-<div class="row content-middle">
-    <span class="greedy push-next">Usage History</span>
-    <toggle-filter class="duration-chooser" params="
-        options: durationOptions,
-        selected: selectedDuration
-    "></toggle-filter>
-</div>
-
-<chart-legend class="push-next" params="items: legendItems"></chart-legend>
-
-<chartjs class="greedy push-both" params="
-    type: 'line',
-    options: chartOptions,
-    data: chartData
-"></chartjs>
