+++ frontend/src/app/components/overview/system-health/system-health.html
@@ -1,0 +1,8 @@
+<div class="pad borders greedy">
+    <h2 class="heading3 row content-middle">
+        <span class=" greedy">System Health</span>
+
+        <svg-icon class="icon-small" params="name: 'question'"></svg-icon>
+    </h2>
+    <hr>
+</div>
