+++ frontend/src/app/components/overview/system-health/system-health.less
@@ -1,3 +1,3 @@
-cluster-summary {
+system-health {
     display: block;
 }
