+++ frontend/src/app/components/register.js
@@ -44,10 +44,11 @@ export default function register(ko) {
     ko.components.register('install-node-modal',    require('./overview/install-node-modal/install-node-modal').default);
     ko.components.register('connect-app-wizard',    require('./overview/connect-app-wizard/connect-app-wizard').default);
     ko.components.register('after-upgrade-modal',   require('./overview/after-upgrade-modal/after-upgrade-modal').default);
-    ko.components.register('storage-usage-summary', require('./overview/storage-usage-summary/storage-usage-summary').default);
-    ko.components.register('nodes-summary', require('./overview/nodes-summary/nodes-summary').default);
-    ko.components.register('cloud-summary', require('./overview/cloud-summary/cloud-summary').default);
-    ko.components.register('cluster-summary', require('./overview/cluster-summary/cluster-summary').default);
+    ko.components.register('buckets-overview',      require('./overview/buckets-overview/buckets-overview').default);
+    ko.components.register('pools-overview',        require('./overview/pools-overview/pools-overview').default);
+    ko.components.register('cloud-overview',        require('./overview/cloud-overview/cloud-overview').default);
+    ko.components.register('cluster-overview',      require('./overview/cluster-overview/cluster-overview').default);
+    ko.components.register('system-health', require('./overview/system-health/system-health').default);
     /** INJECT:overview **/
 
     // -------------------------------
